group the poop
moves old code out of controller and into buildjob pkg since that's
where it will be used.
