--- controllers/util.go
@@ -1,30 +1,0 @@
-package controllers
-
-import (
-	"fmt"
-	"strings"
-
-	"github.com/go-logr/logr"
-	"github.com/pkg/errors"
-)
-
-// These are not informative errors and are captured by the progressui display in a better way
-var ignoredErrors = map[string]interface{}{
-	"runc did not terminate successfully": nil,
-}
-
-// This logs the underlying error from a build when the display channels inside builder.embedded have not yet been initialized
-// or the error comes after the embedded driver has been run (e.g. image size limit has been hit)
-func logError(log logr.Logger, err error) {
-	if unwrappedError := errors.Unwrap(err); unwrappedError != nil {
-		err = unwrappedError
-	}
-
-	cause := errors.Cause(err)
-	if _, ok := ignoredErrors[cause.Error()]; ok {
-		return
-	}
-
-	log.Info(strings.Repeat("=", 70))
-	log.Info(fmt.Sprintf("Error during image build and push: %s", cause.Error()))
-}
